// m-hub-frontend/src/app/enums/mdab-material-group.enum.ts
/**
 * Enum for supported material groups.
 */

export enum MdabMaterialGroup {
    mg_1 = 'Beton inkl. Estrich',
    mg_2 = 'Ziegelmauerwerk',
    mg_3 = 'Holz',
    mg_4 = 'Stahl',
    mg_5 = 'Glas',
    mg_6 = 'Mineralwolle',
    mg_7 = 'Polystyrol',
    mg_8 = 'Gipsplatten, Gips',
    mg_9 = 'Sand, Kies, Stein',
  } 
  